+++ frontend/src/main.ts
@@ -1,4 +1,4 @@
-import init, { DaNode, initialize, NodeStatus } from "@danu/danu";
+import init, { DaNode, initialize, NodeStatus, RealmID } from "@danu/danu";
 
 // DOM elements
 const connectionStatus = document.getElementById(
@@ -38,31 +38,23 @@ async function initApp() {
     addLog("Setting up logging and panic hooks...", "info");
     initialize();
 
-    addLog("Creating FlNode instance...", "info");
+    addLog("Creating DaNode instance...", "info");
     node = await DaNode.from_default();
-
-    // Set up event listener to receive events from the package
     node.set_event_listener(new_event);
-    /*
-      dhtConnections.textContent = status.dht_connections.toString();
 
-      const stats = node.get_stats();
-      messagesSent.textContent = stats.messages_sent.toString();
-      messagesReceived.textContent = stats.messages_received.toString();
-      uptime.textContent = `${stats.uptime_seconds}s`;
- */
-
-    addLog("Fledger Browser initialized successfully", "success");
+    addLog("Danu Browser initialized successfully", "success");
   } catch (error) {
     addLog(`Initialization error: ${error}`, "error");
     console.error("Initialization error:", error);
   }
 }
 
-function new_event(status: NodeStatus, args: any) {
-  console.log(`Handling Event: ${status} - args: ${args}`);
+const realms: RealmID[] = [];
+
+async function new_event(status: NodeStatus, args: any) {
   switch (status) {
     case NodeStatus.ConnectSignal:
+      addLog("Connected to Signal");
       connectionStatus.textContent = "Connected";
       connectionStatus.className = `status-value connected`;
       break;
@@ -71,16 +63,17 @@ function new_event(status: NodeStatus, args: any) {
       connectionStatus.className = `status-value disconnected`;
       break;
     case NodeStatus.ConnectedNodes:
+      addLog(`Connected to ${args} nodes`);
       dhtConnections.textContent = args.toString();
       break;
     case NodeStatus.AvailableNodes:
+      addLog(`${args} nodes available`);
       onlineNodes.textContent = args.toString();
       break;
     case NodeStatus.RealmAvailable:
-      console.log("Got realms:");
-      for (const realm of args) {
-        console.log(realm);
-      }
+      addLog(`${args.length} realms available`);
+      realms.splice(0);
+      realms.push(...args);
   }
 }